backend/requirements.txt
pyserial
requests
numpy
pandas
scikit-learn
fastapi
uvicorn
pydantic
websockets
pydantic-settings